Combine operations.rs and helpers.rs
